+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from './../../models/product.model';
 import { ProductsService } from './../../services/products.service';
+import { CommunicatorService } from './../../../shared/services/communicator.service';
 
 @Component({
   selector: 'app-product-list',
@@ -10,7 +11,7 @@ import { ProductsService } from './../../services/products.service';
 export class ProductListComponent implements OnInit {
   products: Array<Product>;
   productsInBasket: Array<Product>;
-  constructor(private productService: ProductsService) { }
+  constructor(private productService: ProductsService, private communicatorService: CommunicatorService) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
@@ -22,6 +23,8 @@ export class ProductListComponent implements OnInit {
   onAddBasket(product: Product): void {
     console.log('add basket event ProductListComponent ', product);
     this.productsInBasket.push(product);
+    this.communicatorService.publishProduct(product);
+
     console.log('products count in basket ', this.productsInBasket.length);
   }