+++ changelog.md
@@ -10,3 +10,4 @@
 8. Add cart-item component
 9. Add cart-list component
 10. Add cart item delete and quantaty change functionality
+11. With help of communcator service, pass information about adding of product to the cart between product-list and cart-list components